Add MIME type and parser method to bitcode decoder
To make space for the bitcode decoder a new MIME type and file extension are needed. To remove ambiguity the text based IR is renamed LLVM_IR and the bitcode is LLVM_BITCODE, Sulong libraries are unchanged. The parser result abstraction is now used and the parseFile has been renamed to reflect its narrowed use.