Add support for bitcode decoding to Sulong
LLVM has been modified to parse both IR and bitcode. The bitcode decoding is handled from the llvm.frontend project. It is hoped that the IR parser and factories will be unified to remove duplication. At the moment this is not possible becuase the abstraction uses specialised classes e.g. types. At the moment the library in the background is not public so the suite has not yet been updated to reflect the changes.