Move most parts of LLVMWriter into the different types
This means toString() of instructions now return their LLVM IR 3.2 equivalient.
We have to think how to implement the differences beweeen various LLVM IR Versions,
but for now, this apporach is pretty clean and should be easy extendable for the future.

Because Symbols are currently not named inside InstructionGeneratorFacade, output
of a InstructionBlock made with this Facade doesn't make much sense. (no variable names)